+++ src/features/import/components/Importer/Configure/Mapping/index.tsx
@@ -1,18 +1,12 @@
 import { FC } from 'react';
 import { Box, Divider, Typography } from '@mui/material';
 
-import globalMessageIds from 'core/i18n/globalMessageIds';
 import messageIds from 'features/import/l10n/messageIds';
-import { NATIVE_PERSON_FIELDS } from 'features/views/components/types';
 import range from 'utils/range';
-import useCustomFields from 'features/profile/hooks/useCustomFields';
+import useFields from 'features/import/hooks/useFields';
 import { useMessages } from 'core/i18n';
 import { useNumericRouteParams } from 'core/hooks';
-import MappingRow, {
-  ExperimentalFieldTypes,
-  ExperimentColumn,
-  ExperimentField,
-} from './MappingRow';
+import MappingRow, { ExperimentColumn } from './MappingRow';
 
 export interface ExperimentRow {
   data: (string | number | null)[];
@@ -26,58 +20,6 @@ interface MappingProps {
   rows?: ExperimentRow[];
   selectedColumns: number[];
 }
-
-const useFields = (orgId: number): ExperimentField[] => {
-  const globalMessages = useMessages(globalMessageIds);
-  const messages = useMessages(messageIds);
-  const customFields = useCustomFields(orgId).data ?? [];
-
-  const fieldsWithoutId: { slug: string; title: string }[] = [];
-
-  customFields.forEach((field) =>
-    fieldsWithoutId.push({
-      slug: field.slug,
-      title: field.title,
-    })
-  );
-
-  Object.values(NATIVE_PERSON_FIELDS).forEach((fieldSlug) =>
-    fieldsWithoutId.push({
-      slug: fieldSlug,
-      title: globalMessages.personFields[fieldSlug](),
-    })
-  );
-
-  const fields = fieldsWithoutId
-    .filter((field) => field.slug != 'id' && field.slug != 'ext_id')
-    .map((field) => ({
-      ...field,
-      type: ExperimentalFieldTypes.BASIC,
-    }));
-
-  fields.push({
-    slug: 'org',
-    title: messages.configuration.mapping.organization(),
-    type: ExperimentalFieldTypes.ORGANIZATION,
-  });
-  fields.push({
-    slug: 'tags',
-    title: messages.configuration.mapping.tags(),
-    type: ExperimentalFieldTypes.TAG,
-  });
-  fields.push({
-    slug: 'id',
-    title: messages.configuration.mapping.id(),
-    type: ExperimentalFieldTypes.ID,
-  });
-
-  return fields
-    .map((field, index) => ({
-      ...field,
-      id: index + 1,
-    }))
-    .sort((field1, field2) => field1.title.localeCompare(field2.title));
-};
 
 const Mapping: FC<MappingProps> = ({
   currentlyMapping,
